Core: Added proper support for "offline" signatures.
The main difference against the online signatures is that we're using predefined "offline" key for APP_SECRET & APP_KEY constants. This is mostly because it's problematic to match the right APP_SECRET during the validation on the server. In the offline mode, the server has just the decimalized signature and that's not enough for proper match.